Finished introducing constant_zero, constant_one, isZero() and isOne() to carl.

+++ src/carl/numbers/constants.h
@@ -38,7 +38,7 @@ namespace carl
 	struct constant_zero {
 		static const T& get() 
 		{
-			static T t(0);
+			static T t((int)0);
 			return t;
 		}
 	};
@@ -47,7 +47,7 @@ namespace carl
 	struct constant_one {
 		static const T& get() 
 		{
-			static T t(1);
+			static T t((int)1);
 			return t;
 		}
 	};
